# src/main/java/com/anonymous/wall/admin/controller/AdminCommentController.java
package com.anonymous.wall.admin.controller;

import com.anonymous.wall.admin.service.AdminCommentService;
import com.anonymous.wall.entity.Comment;
import com.anonymous.wall.model.AdminCommentDTO;
import io.micronaut.core.annotation.Nullable;
import io.micronaut.data.model.Page;
import io.micronaut.data.model.Pageable;
import io.micronaut.http.HttpResponse;
import io.micronaut.http.HttpRequest;
import io.micronaut.http.annotation.*;
import io.micronaut.security.annotation.Secured;
import jakarta.inject.Inject;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import java.util.HashMap;
import java.util.List;
import java.util.Map;
import java.util.UUID;
import java.util.stream.Collectors;

/**
 * Admin controller for comment moderation
 */
@Controller("/api/v1/admin/comments")
public class AdminCommentController {
    
    private static final Logger log = LoggerFactory.getLogger(AdminCommentController.class);
    
    @Inject
    private AdminCommentService adminCommentService;
    
    private AdminCommentDTO mapCommentToDTO(Comment comment) {
        AdminCommentDTO dto = new AdminCommentDTO();
        dto.setId(comment.getId());
        dto.setPostId(comment.getParentId());
        dto.setUserId(comment.getUserId());
        dto.setProfileName(comment.getProfileName());
        dto.setText(comment.getText());
        dto.setHidden(comment.isHidden());
        dto.setCreatedAt(comment.getCreatedAt());
        return dto;
    }
    
    @Get
    @Secured({"ADMIN", "MODERATOR"})
    public HttpResponse<Object> getAllComments(
            @QueryValue(defaultValue = "1") int page,
            @QueryValue(defaultValue = "20") int limit,
            @Nullable @QueryValue String userId,
            @Nullable @QueryValue Boolean hidden,
            @Nullable @QueryValue String sortBy,
            @Nullable @QueryValue String sortOrder,
            HttpRequest<?> request) {
        
        if (page < 1) page = 1;
        if (limit < 1 || limit > 100) limit = 20;
        
        Pageable pageable = Pageable.from(page - 1, limit);
        UUID userIdUuid = userId != null ? UUID.fromString(userId) : null;
        Page<Comment> commentsPage = adminCommentService.getAllComments(pageable, userIdUuid, hidden, sortBy, sortOrder);
        
        List<AdminCommentDTO> commentDTOs = commentsPage.getContent().stream()
                .map(this::mapCommentToDTO)
                .collect(Collectors.toList());
        
        Map<String, Object> response = new HashMap<>();
        response.put("data", commentDTOs);
        
        Map<String, Object> pagination = new HashMap<>();
        pagination.put("page", page);
        pagination.put("limit", limit);
        pagination.put("total", commentsPage.getTotalSize());
        pagination.put("totalPages", commentsPage.getTotalPages());
        response.put("pagination", pagination);
        
        return HttpResponse.ok(response);
    }
    
    @Get("/{id}")
    @Secured({"ADMIN", "MODERATOR"})
    public HttpResponse<AdminCommentDTO> getCommentById(@PathVariable String id) {
        UUID commentId = UUID.fromString(id);
        Comment comment = adminCommentService.getCommentById(commentId);
        return HttpResponse.ok(mapCommentToDTO(comment));
    }
    
    @Delete("/{id}")
    @Secured({"ADMIN", "MODERATOR"})
    public HttpResponse<Object> deleteComment(@PathVariable String id) {
        log.info("Admin deleting comment: {}", id);
        UUID commentId = UUID.fromString(id);
        adminCommentService.deleteComment(commentId);
        Map<String, String> response = new HashMap<>();
        response.put("message", "Comment deleted successfully");
        return HttpResponse.ok(response);
    }

    @Put("/{id}/hide")
    @Secured({"ADMIN", "MODERATOR"})
    public HttpResponse<Object> hideComment(@PathVariable String id) {
        log.info("Admin hiding comment: {}", id);
        UUID commentId = UUID.fromString(id);
        adminCommentService.hideComment(commentId);
        Map<String, String> response = new HashMap<>();
        response.put("message", "Comment hidden successfully");
        return HttpResponse.ok(response);
    }

    @Put("/{id}/unhide")
    @Secured({"ADMIN", "MODERATOR"})
    public HttpResponse<Object> unhideComment(@PathVariable String id) {
        log.info("Admin unhiding comment: {}", id);
        UUID commentId = UUID.fromString(id);
        adminCommentService.unhideComment(commentId);
        Map<String, String> response = new HashMap<>();
        response.put("message", "Comment unhidden successfully");
        return HttpResponse.ok(response);
    }
}