# src/main/resources/api.yml
openapi: 3.0.0
info:
  title: Campus Wall API
  version: "1.0.0"
  description: API for campus social features
servers:
  - url: /api/v1
...
paths:
...
  /posts/{postId}/comments/{commentId}/hide:
    patch:
      tags: [posts]
      summary: Hide a comment
      description: Hide a comment (soft-delete). Only the comment author can hide their own comment.
      operationId: hideComment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: string
          description: ID of the post
        - in: path
          name: commentId
          required: true
          schema:
            type: string
          description: ID of the comment to hide
      responses:
        "200":
          description: Comment hidden successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comment hidden successfully"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - only comment author can hide their own comment
        "404":
          description: Post or comment not found
  /posts/{postId}/comments/{commentId}/unhide:
    patch:
      tags: [posts]
      summary: Unhide a comment
      description: Unhide a previously hidden comment (undo soft-delete). Only the comment author can unhide their own comment.
      operationId: unhideComment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: string
          description: ID of the post
        - in: path
          name: commentId
          required: true
          schema:
            type: string
          description: ID of the comment to unhide
      responses:
        "200":
          description: Comment unhidden successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comment unhidden successfully"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - only comment author can unhide their own comment
        "404":
          description: Post or comment not found
...
  /internships/{internshipId}:
    get:
      tags: [internships]
      summary: Get internship posting by ID
      description: Retrieve a specific internship posting by its ID
      operationId: getInternship
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: internshipId
          required: true
          schema:
            type: string
          description: ID of the internship posting to retrieve
      responses:
        "200":
          description: Internship posting retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternshipDTO"
        "404":
          description: Internship posting not found
        "401":
          description: Unauthorized - authentication required
  /internships/{internshipId}/hide:
    patch:
      tags: [internships]
      summary: Hide an internship posting
      description: Hide an internship posting (soft-delete). Only the posting author can hide their own posting.
      operationId: hideInternship
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: internshipId
          required: true
          schema:
            type: string
          description: ID of the internship posting to hide
      responses:
        "200":
          description: Internship posting hidden successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          description: Internship posting not found
        "403":
          description: Forbidden - You can only hide your own postings
        "401":
          description: Unauthorized - authentication required
  /internships/{internshipId}/unhide:
    patch:
      tags: [internships]
      summary: Unhide an internship posting
      description: Unhide a previously hidden internship posting. Only the posting author can unhide their own posting.
      operationId: unhideInternship
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: internshipId
          required: true
          schema:
            type: string
          description: ID of the internship posting to unhide
      responses:
        "200":
          description: Internship posting unhidden successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          description: Internship posting not found
        "403":
          description: Forbidden - You can only unhide your own postings
        "401":
          description: Unauthorized - authentication required
...
  /admin/comments:
    get:
      tags: [admin]
      summary: Get all comments
      description: Admin endpoint to retrieve all comments with pagination, filters, and sorting
      operationId: adminGetComments
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - in: query
          name: userId
          required: false
          schema:
            type: string
            format: uuid
          description: Filter comments by user ID
        - in: query
          name: hidden
          required: false
          schema:
            type: boolean
          description: Filter by hidden status
        - in: query
          name: sortBy
          required: false
          schema:
            type: string
            enum: [createdAt, userId]
          description: Field to sort by
        - in: query
          name: sortOrder
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order (ascending or descending)
      responses:
        "200":
          description: Comments retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminCommentDTO"
                  pagination:
                    $ref: "#/components/schemas/PaginationInfo"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - admin access required
  /admin/comments/{id}:
    get:
      tags: [admin]
      summary: Get a comment by ID
      description: Admin endpoint to retrieve a specific comment by its ID
      operationId: adminGetCommentById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Comment ID to retrieve
      responses:
        "200":
          description: Comment retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCommentDTO"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - admin access required
        "404":
          description: Comment not found
    delete:
      tags: [admin]
      summary: Delete a comment
      description: Admin endpoint to soft-delete (hide) a comment
      operationId: adminDeleteComment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Comment ID to delete
      responses:
        "200":
          description: Comment deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comment deleted successfully"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - admin access required
        "404":
          description: Comment not found
  /admin/comments/{id}/hide:
    put:
      tags: [admin]
      summary: Hide a comment
      operationId: adminHideComment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Comment hidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/comments/{id}/unhide:
    put:
      tags: [admin]
      summary: Unhide a comment
      operationId: adminUnhideComment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Comment unhidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
  /admin/internships:
    get:
      tags: [admin]
      summary: Get all internships
      operationId: adminGetInternships
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
        - in: query
          name: hidden
          schema:
            type: boolean
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [createdAt, commentCount]
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Internships retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminInternshipDTO"
                  pagination:
                    $ref: "#/components/schemas/PaginationInfo"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/internships/{id}:
    get:
      tags: [admin]
      summary: Get an internship by ID
      operationId: adminGetInternshipById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Internship retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInternshipDTO"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/internships/{id}/hide:
    put:
      tags: [admin]
      summary: Hide an internship
      operationId: adminHideInternship
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Internship hidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/internships/{id}/unhide:
    put:
      tags: [admin]
      summary: Unhide an internship
      operationId: adminUnhideInternship
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Internship unhidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
components:
  schemas:
...
    InternshipDTO:
      type: object
      properties:
        id:
          type: string
        company:
          type: string
        role:
          type: string
        salary:
          type: string
        location:
          type: string
        description:
          type: string
        deadline:
          type: string
          format: date
        wall:
          type: string
          enum: [campus, national]
          description: Which wall the internship is on
        comments:
          type: integer
          description: Number of comments
        author:
          type: object
          properties:
            id:
              type: string
            profileName:
              type: string
              description: Profile name of the posting author
            isAnonymous:
              type: boolean
              description: Whether author is anonymous
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ----- Marketplace -----
...
    AdminCommentDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        postId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        profileName:
          type: string
        text:
          type: string
        hidden:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    AdminInternshipDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        profileName:
          type: string
        company:
          type: string
        role:
          type: string
        salary:
          type: string
        location:
          type: string
        description:
          type: string
        deadline:
          type: string
          format: date
        wall:
          type: string
          enum: [campus, national]
        schoolDomain:
          type: string
        commentCount:
          type: integer
        hidden:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Number of items per page
        total:
          type: integer
          description: Total number of items
        totalPages:
          type: integer
          description: Total number of pages
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT